[Nicolo] networking

+++ networking/number-system-conversion/DecimalConversion/DecimalConversion.h
@@ -4,7 +4,7 @@
 class DecimalConversion {
 	
 	public:
-		virtual long long int	decimalToBaseN (int decimal, int n);
+		virtual int	decimalToBaseN (int decimal, int n);
 		virtual std::string decimalToHex (int decimal);
 	
 };
